pyproject.toml
[project]
name = "openclaw-demo"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langchain",
    "langchain-openai",
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "streamlit",
    "websockets",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "requests>=2.32.5",
    "aiosqlite>=0.22.1",
    "python-telegram-bot>=22.6",
]

